requirements.txt
pandas>=2.0
numpy>=1.26
pyarrow>=14.0
aws-cdk-lib>=2.100
constructs>=10.0
boto3>=1.28
